Done Wishlist User Specific Data

+++ src/components/Frequent/FrequentlyPurchased.jsx
@@ -60,8 +60,7 @@ const FrequentlyPurchased = () => {
           });
           setLoggedInUser(user);
         } else {
-          setLoggedInUser(null);
-          navigate("/login-user");
+          setLoggedInUser(null);          
         }
       });
 
@@ -127,39 +126,47 @@ const FrequentlyPurchased = () => {
   }
 
   const toggleWishlist = async (productName, productRating, productTotalRating, productDescription, productOfferPrice, productPrice, productOff, productId, id, index) => {    
-    setCurrentIndex(index)
-    setSpinner(true)
-    let checkInDB = false   
-    
-    const getAllDoc = await getDocs(wishlistCollection);
 
-    getAllDoc.forEach((doc) => {       
-        if(productName === doc.data().productName) {
-            checkInDB = true                        
-        }
-    })   
-    
-    let userId = userDetails.userId
-
-    if(!checkInDB) {
-        const querySnapshot = await addDoc(wishlistCollection, {productName, productRating, productTotalRating, productDescription, productOfferPrice, productPrice, productOff, productId, id, userId})  
-
-        setSpinner(false)
-        showWishlistMessage(1);
+    if(checkLoggedInUser) {
+      setCurrentIndex(index)
+      setSpinner(true)
+      let checkInDB = false   
+      
+      const getAllDoc = await getDocs(wishlistCollection);
+  
+      getAllDoc.forEach((doc) => {       
+          if(productName === doc.data().productName) {
+              checkInDB = true                        
+          }
+      })   
+      
+      let userId = userDetails.userId
+  
+      if(!checkInDB) {
+          const querySnapshot = await addDoc(wishlistCollection, {productName, productRating, productTotalRating, productDescription, productOfferPrice, productPrice, productOff, productId, id, userId})  
+  
+          setSpinner(false)
+          showWishlistMessage(1);
+      }
+  
+      else {
+          const deleteFromWishlist = await getDocs(
+              query(collection(db, "/wishlist"), where("userId", "==", userId), where("productName", "==", productName))
+          );
+         
+          deleteFromWishlist.forEach((doc) => {   
+              deleteDoc(doc.ref);
+            }); 
+          
+          setSpinner(false)
+          showWishlistMessage(0); 
+      }
     }
 
     else {
-        const deleteFromWishlist = await getDocs(
-            query(collection(db, "/wishlist"), where("productName", "==", productName))
-        );
-       
-        deleteFromWishlist.forEach((doc) => {   
-            deleteDoc(doc.ref);
-          }); 
-        
-        setSpinner(false)
-        showWishlistMessage(0); 
+      navigate('/login-user')
     }
+    
   }
 
   return (
